+++ src/routes/index.ts
@@ -10,6 +10,7 @@ export * from "./BaseCalendarShareLinkRoute.js";
 export * from "./BaseContactRoute.js";
 export * from "./BaseDataExportRoute.js";
 export * from "./BaseDataSubjectErasureRequestRoute.js";
+export * from "./BaseDirectoryRoute.js";
 export * from "./BaseDistributionListRoute.js";
 export * from "./BaseDomainRoute.js";
 export * from "./BaseEncryptionPolicyRoute.js";