+++ src/routes/mongo/index.ts
@@ -7,6 +7,7 @@ export * from "./ContactListRouteMongo.js";
 export * from "./ContactRouteMongo.js";
 export * from "./DataExportRequestRouteMongo.js";
 export * from "./DataSubjectErasureRequestRouteMongo.js";
+export * from "./DirectoryRouteMongo.js";
 export * from "./DistributionListRouteMongo.js";
 export * from "./DomainRouteMongo.js";
 export * from "./EncryptionPolicyRouteMongo.js";