+++ src/routes/sql/index.ts
@@ -7,6 +7,7 @@ export * from "./ContactListRouteSQL.js";
 export * from "./ContactRouteSQL.js";
 export * from "./DataExportRequestRouteSQL.js";
 export * from "./DataSubjectErasureRequestRouteSQL.js";
+export * from "./DirectoryRouteSQL.js";
 export * from "./DistributionListRouteSQL.js";
 export * from "./DomainRouteSQL.js";
 export * from "./EncryptionPolicyRouteSQL.js";