+++ test/server-mongo/routes/DirectoryRoute.ts
@@ -1,0 +1,6 @@
+import { RouteDecorators } from "@rapidrest/service-core";
+import { DirectoryRouteMongo } from "../../../src/routes/mongo/DirectoryRouteMongo.js";
+const { Route } = RouteDecorators;
+
+@Route("/mongo/directory")
+export class DirectoryRoute extends DirectoryRouteMongo {}