+++ test/server-sql/routes/DirectoryRoute.ts
@@ -1,0 +1,6 @@
+import { RouteDecorators } from "@rapidrest/service-core";
+import { DirectoryRouteSQL } from "../../../src/routes/sql/DirectoryRouteSQL.js";
+const { Route } = RouteDecorators;
+
+@Route("/sql/directory")
+export class DirectoryRoute extends DirectoryRouteSQL {}